refactor pv & write echo example
