+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "packetvisor"
+name = "pv"
 version = "0.1.0"
 edition = "2021"
